refactor and add read_csv
