+++ python/Cargo.toml
@@ -29,8 +29,11 @@ rust-version = "1.64"
 include = ["/src", "/pyballista", "/LICENSE.txt", "pyproject.toml", "Cargo.toml", "Cargo.lock"]
 
 [dependencies]
+async-trait = "0.1.77"
 ballista = { path = "../ballista/client", version = "0.12.0" }
+ballista-core = { path = "../ballista/core", version = "0.12.0" }
 datafusion = "35.0.0"
+datafusion-proto = "35.0.0"
 datafusion-python = "35.0.0"
 pyo3 = { version = "0.20", features = ["extension-module", "abi3", "abi3-py38"] }
 tokio = { version = "1.35", features = ["macros", "rt", "rt-multi-thread", "sync"] }
