+++ python/src/dataframe.rs
@@ -1,0 +1,51 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+use crate::utils::wait_for_future;
+use datafusion::arrow::pyarrow::ToPyArrow;
+use datafusion::prelude::DataFrame;
+use pyo3::prelude::*;
+use std::sync::Arc;
+
+/// Ballista DataFrame is a wrapper around a DataFusion
+/// DataFrame. Do we even need this?
+#[pyclass(name = "DataFrame", module = "pyballista", subclass)]
+#[derive(Clone)]
+pub struct PyDataFrame {
+    /// DataFusion DataFrame
+    df: Arc<DataFrame>,
+}
+
+impl PyDataFrame {
+    /// creates a new PyDataFrame
+    pub fn new(df: DataFrame) -> Self {
+        Self { df: Arc::new(df) }
+    }
+}
+
+#[pymethods]
+impl PyDataFrame {
+    /// Executes the plan, returning a list of `RecordBatch`es.
+    /// Unless some order is specified in the plan, there is no
+    /// guarantee of the order of the result.
+    fn collect(&self, py: Python) -> PyResult<Vec<PyObject>> {
+        let batches = wait_for_future(py, self.df.as_ref().clone().collect())?;
+        // cannot use PyResult<Vec<RecordBatch>> return type due to
+        // https://github.com/PyO3/pyo3/issues/1813
+        batches.into_iter().map(|rb| rb.to_pyarrow(py)).collect()
+    }
+}
